refactor(init): split init flows into separate files
Extract 9 private methods from the monolithic InitCommand class into
5 standalone modules under actions/init/, preparing the file structure
for the bundled-create branch to cleanly rebase on top.

New files:
- initHelpers.ts (shouldPrompt, flagOrDefault, getPostInitMCPPrompt, writeStagingEnvIfNeeded)
- scaffoldTemplate.ts (selectTemplate, scaffoldAndInstall)
- initNextJs.ts (initNextJs, writeOrOverwrite, writeSourceFiles)
- initApp.ts (initApp — app template flow)
- initStudio.ts (initStudio — studio template flow)
